File modes: h5py-like file permissions
Previously, z5py.File objects could only be opened in a mode analogous
to h5py's 'a'. However, other modes are useful.

Because this API is based on h5py, "create" permissions refer only to
the top-level File object (represented as a directory). Objects opened
with only a "write" permission can still create directories and OS
files, because they are not considered z5py.Files.