Implement file permissions in c++ and python
The file modes / permissions are implemented in include\z5\hadnle\file_mode.hxx
and this class is exposed to python.
In c++, the file handles carry the permissions object, in python it is a member
of the base class.
Exceptions are not consistent yet! (C++ throws runtime errors, python system erros).